+++ examples/tutorial/02-two-region-model/emissions.jl
@@ -1,20 +1,20 @@
-using Mimi                                          #Make sure to call Mimi again
+using Mimi                                          # Make sure to call Mimi again
 
 @defcomp emissions begin
-    regions     = Index()                           #The regions index must be specified for each component
+    regions     = Index()                           # The regions index must be specified for each component
 
-    E           = Variable(index=[time, regions])   #Total greenhouse gas emissions
-    E_Global    = Variable(index=[time])            #Global emissions (sum of regional emissions)
-    sigma       = Parameter(index=[time, regions])  #Emissions output ratio
-    YGROSS      = Parameter(index=[time, regions])  #Gross output - Note that YGROSS is now a parameter
+    E           = Variable(index=[time, regions])   # Total greenhouse gas emissions
+    E_Global    = Variable(index=[time])            # Global emissions (sum of regional emissions)
+    sigma       = Parameter(index=[time, regions])  # Emissions output ratio
+    YGROSS      = Parameter(index=[time, regions])  # Gross output - Note that YGROSS is now a parameter
 
     function run(p, v, d, t)
-        #Define an eqation for E
+        # Define an equation for E
         for r in d.regions
             v.E[t,r] = p.YGROSS[t,r] * p.sigma[t,r]
         end
 
-        #Define an equation for E_Global
+        # Define an equation for E_Global
         for r in d.regions
             v.E_Global[t] = sum(v.E[t,:])
         end