+++ src/Mimi.jl
@@ -54,8 +54,8 @@ export
     indexcount,
     indexlabels,
     indexvalues,
-    is_final_timestep,
-    is_first_timestep,
+    is_stop,
+    is_start,
     load_comps,
     modeldef,
     name,
@@ -86,10 +86,13 @@ include("core/connections.jl")
 include("core/defs.jl")
 include("core/defcomp.jl")
 include("core/instances.jl")
-# include("core/mimi-core.jl")
 include("core/references.jl")
 include("core/time.jl")
 include("core/model.jl")
+
+include("mcs/mcs_types.jl")
+include("mcs/EmpiricalDistribution.jl")
+include("mcs/montecarlo.jl")
 
 include("utils/graph.jl")
 # include("utils/plotting.jl")